Creature Save Data Update
I made it so that the creatures now have their data stored in an array. This will allow the creatures to be saved for each day instead of making a new one every day. This will make it so that the creatures will lose hunger after every day and once they get down to 0, they will die.