Gather More Food Function
Attempted to put in a gatherFood() function, though it is not working correctly. Now things are going crazy and I don't understand why. lol. Anyways, the project is basically finished already, I am just adding more stuff for fun at this point. It is starting to get increasingly more complicated.